Refactor various types into better hierarchy
