+++ data/factory/standard.go
@@ -19,12 +19,14 @@ import (
 	"github.com/tidepool-org/platform/data/types/base/basal/scheduled"
 	"github.com/tidepool-org/platform/data/types/base/basal/suspend"
 	"github.com/tidepool-org/platform/data/types/base/basal/temporary"
+	"github.com/tidepool-org/platform/data/types/base/blood/glucose/continuous"
+	"github.com/tidepool-org/platform/data/types/base/blood/glucose/selfmonitored"
+	"github.com/tidepool-org/platform/data/types/base/blood/ketone"
 	"github.com/tidepool-org/platform/data/types/base/bolus"
 	"github.com/tidepool-org/platform/data/types/base/bolus/calculator"
 	"github.com/tidepool-org/platform/data/types/base/bolus/combination"
 	"github.com/tidepool-org/platform/data/types/base/bolus/extended"
 	"github.com/tidepool-org/platform/data/types/base/bolus/normal"
-	"github.com/tidepool-org/platform/data/types/base/continuous"
 	"github.com/tidepool-org/platform/data/types/base/device"
 	"github.com/tidepool-org/platform/data/types/base/device/alarm"
 	"github.com/tidepool-org/platform/data/types/base/device/calibration"
@@ -32,9 +34,7 @@ import (
 	"github.com/tidepool-org/platform/data/types/base/device/reservoirchange"
 	"github.com/tidepool-org/platform/data/types/base/device/status"
 	"github.com/tidepool-org/platform/data/types/base/device/timechange"
-	"github.com/tidepool-org/platform/data/types/base/ketone"
-	"github.com/tidepool-org/platform/data/types/base/pump"
-	"github.com/tidepool-org/platform/data/types/base/selfmonitored"
+	"github.com/tidepool-org/platform/data/types/base/settings/pump"
 	"github.com/tidepool-org/platform/data/types/base/upload"
 )
