+++ data/types/base/blood/glucose/continuous/continuous.go
