+++ data/types/base/blood/glucose/selfmonitored/self_monitored.go
